+++ qiskit/extensions/standard/cz.py
@@ -16,6 +16,7 @@ from qiskit import QuantumCircuit
 from qiskit._instructionset import InstructionSet
 from qiskit._quantumregister import QuantumRegister
 from qiskit.extensions.standard import header  # pylint: disable=unused-import
+from qiskit.extensions.standard.x import XGate
 
 
 class CzGate(Gate):
@@ -39,6 +40,25 @@ class CzGate(Gate):
     def reapply(self, circ):
         """Reapply this gate to corresponding qubits in circ."""
         self._modifiers(circ.cz(self.arg[0], self.arg[1]))
+
+    def q_if(self, *qregs):
+        if not qregs:
+            return self
+
+        # Dynamically create the Gate that will perform
+        # a multi-controlled Z gate by using a multi-controlled
+        # X gate and 2 H gates.
+        class _MultiControlledZGateWithCx(CompositeGate):
+            def __init__(self, circuit, target, *controls):
+                super().__init__(self.__class__.__name__,  # name
+                                 [],                       # parameters
+                                 [target] + controls,      # qubits
+                                 circuit)                  # circuit
+                self.h(target)
+                self._attach(XGate(target, self).q_if(*controls))
+                self.h(target)
+
+        return _MultiControlledZGateWithCx(self._get_circuit(), self.arg[1], [self.arg[1]] + qregs)
 
 
 def cz(self, ctl, tgt):