+++ qiskit/extensions/standard/h.py
@@ -16,6 +16,7 @@ from qiskit import InstructionSet
 from qiskit import QuantumCircuit
 from qiskit import QuantumRegister
 from qiskit.extensions.standard import header  # pylint: disable=unused-import
+from qiskit.extensions.standard.ch import CHGate
 
 
 class HGate(Gate):
@@ -37,6 +38,12 @@ class HGate(Gate):
     def reapply(self, circ):
         """Reapply this gate to corresponding qubits in circ."""
         self._modifiers(circ.h(self.arg[0]))
+
+    def q_if(self, *qregs):
+        """Add controls to this gate."""
+        if not qregs:
+            return self
+        return CHGate(qregs[0], self.arg[0], self._get_circuit()).q_if(*qregs[1:])
 
 
 def h(self, q):